refactor(releases): move releases lists outside of the global perspective menu
