+++ packages/sanity/src/core/releases/navbar/GlobalPerspectiveMenu.tsx
@@ -1,24 +1,13 @@
 import {type ReleaseId} from '@sanity/client'
-import {AddIcon, ChevronDownIcon} from '@sanity/icons'
+import {ChevronDownIcon} from '@sanity/icons'
 // eslint-disable-next-line no-restricted-imports -- MenuItem requires props, only supported by @sanity/ui
-import {Box, Button, Flex, Menu, MenuDivider, MenuItem, Spinner} from '@sanity/ui'
-import {useCallback, useMemo, useRef, useState} from 'react'
-import {css, styled} from 'styled-components'
+import {Button, Menu} from '@sanity/ui'
+import {useCallback, useRef, useState} from 'react'
+import {styled} from 'styled-components'
 
 import {MenuButton} from '../../../ui-components'
-import {useTranslation} from '../../i18n'
-import {usePerspective} from '../../perspective/usePerspective'
 import {CreateReleaseDialog} from '../components/dialog/CreateReleaseDialog'
-import {type ReleaseDocument, type ReleaseType} from '../store/types'
-import {useActiveReleases} from '../store/useActiveReleases'
-import {LATEST} from '../util/const'
-import {getReleaseIdFromReleaseDocumentId} from '../util/getReleaseIdFromReleaseDocumentId'
-import {
-  getRangePosition,
-  GlobalPerspectiveMenuItem,
-  type LayerRange,
-} from './GlobalPerspectiveMenuItem'
-import {ReleaseTypeMenuSection} from './ReleaseTypeMenuSection'
+import {ReleasesList} from './ReleasesList'
 import {useScrollIndicatorVisibility} from './useScrollIndicatorVisibility'
 
 const StyledMenu = styled(Menu)`
@@ -26,148 +15,20 @@ const StyledMenu = styled(Menu)`
   max-width: 320px;
 `
 
-const StyledBox = styled(Box)`
-  overflow: auto;
-  max-height: 75vh;
-`
-
-const StyledPublishedBox = styled(Box)<{$removePadding: boolean}>(
-  ({$removePadding}) => css`
-    position: sticky;
-    top: 0;
-    background-color: var(--card-bg-color);
-    z-index: 10;
-    padding-bottom: ${$removePadding ? '4px' : '16px'};
-  `,
-)
-
-const orderedReleaseTypes: ReleaseType[] = ['asap', 'scheduled', 'undecided']
-
-const ASAP_RANGE_OFFSET = 2
-
 export function GlobalPerspectiveMenu({
   selectedReleaseId,
 }: {
   selectedReleaseId: ReleaseId | undefined
 }): React.JSX.Element {
-  const {loading, data: releases} = useActiveReleases()
   const [createBundleDialogOpen, setCreateBundleDialogOpen] = useState(false)
   const styledMenuRef = useRef<HTMLDivElement>(null)
 
   const {isRangeVisible, onScroll, resetRangeVisibility, setScrollContainer, scrollElementRef} =
     useScrollIndicatorVisibility()
 
-  const {t} = useTranslation()
-
-  /* create new release */
-  const handleCreateBundleClick = useCallback(() => {
-    setCreateBundleDialogOpen(true)
-  }, [])
-
   const handleClose = useCallback(() => {
     setCreateBundleDialogOpen(false)
   }, [])
-
-  const sortedReleaseTypeReleases = useMemo(
-    () =>
-      orderedReleaseTypes.reduce<Record<ReleaseType, ReleaseDocument[]>>(
-        (ReleaseTypeReleases, releaseType) => ({
-          ...ReleaseTypeReleases,
-          [releaseType]: releases.filter(({metadata}) => metadata.releaseType === releaseType),
-        }),
-        {} as Record<ReleaseType, ReleaseDocument[]>,
-      ),
-    [releases],
-  )
-
-  const range: LayerRange = useMemo(() => {
-    let lastIndex = 0
-
-    const {asap, scheduled} = sortedReleaseTypeReleases
-    const countAsapReleases = asap.length
-    const countScheduledReleases = scheduled.length
-
-    const offsets = {
-      asap: ASAP_RANGE_OFFSET,
-      scheduled: ASAP_RANGE_OFFSET + countAsapReleases,
-      undecided: ASAP_RANGE_OFFSET + countAsapReleases + countScheduledReleases,
-    }
-
-    const adjustIndexForReleaseType = (type: ReleaseType) => {
-      const groupSubsetReleases = sortedReleaseTypeReleases[type]
-      const offset = offsets[type]
-
-      groupSubsetReleases.forEach((release, groupReleaseIndex) => {
-        const index = offset + groupReleaseIndex
-
-        if (selectedReleaseId === getReleaseIdFromReleaseDocumentId(release._id)) {
-          lastIndex = index
-        }
-      })
-    }
-
-    orderedReleaseTypes.forEach(adjustIndexForReleaseType)
-
-    return {
-      lastIndex,
-      offsets,
-    }
-  }, [selectedReleaseId, sortedReleaseTypeReleases])
-
-  const releasesList = useMemo(() => {
-    if (loading) {
-      return (
-        <Flex padding={4} justify="center" data-testid="spinner">
-          <Spinner muted />
-        </Flex>
-      )
-    }
-
-    return (
-      <Box>
-        <StyledBox ref={setScrollContainer} onScroll={onScroll}>
-          <StyledPublishedBox $removePadding={!releases.length}>
-            <GlobalPerspectiveMenuItem
-              rangePosition={isRangeVisible ? getRangePosition(range, 0) : undefined}
-              release={'published'}
-            />
-            <GlobalPerspectiveMenuItem
-              rangePosition={isRangeVisible ? getRangePosition(range, 1) : undefined}
-              release={LATEST}
-            />
-          </StyledPublishedBox>
-          <>
-            {orderedReleaseTypes.map((releaseType) => (
-              <ReleaseTypeMenuSection
-                key={releaseType}
-                releaseType={releaseType}
-                releases={sortedReleaseTypeReleases[releaseType]}
-                range={range}
-                currentGlobalBundleMenuItemRef={scrollElementRef}
-              />
-            ))}
-          </>
-        </StyledBox>
-        <MenuDivider />
-        <MenuItem
-          icon={AddIcon}
-          onClick={handleCreateBundleClick}
-          text={t('release.action.create-new')}
-        />
-      </Box>
-    )
-  }, [
-    handleCreateBundleClick,
-    isRangeVisible,
-    loading,
-    onScroll,
-    range,
-    releases.length,
-    scrollElementRef,
-    setScrollContainer,
-    sortedReleaseTypeReleases,
-    t,
-  ])
 
   return (
     <>
@@ -186,7 +47,14 @@ export function GlobalPerspectiveMenu({
         onClose={resetRangeVisibility}
         menu={
           <StyledMenu data-testid="release-menu" ref={styledMenuRef}>
-            {releasesList}
+            <ReleasesList
+              setScrollContainer={setScrollContainer}
+              onScroll={onScroll}
+              isRangeVisible={isRangeVisible}
+              scrollElementRef={scrollElementRef}
+              selectedReleaseId={selectedReleaseId}
+              setCreateBundleDialogOpen={setCreateBundleDialogOpen}
+            />
           </StyledMenu>
         }
         popover={{
